Api gateway

+++ transaction/src/main/java/com/example/transaction/services/TransactionServiceImpl.java
@@ -4,13 +4,21 @@ import java.io.IOException;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
+import java.util.Random;
 import java.util.stream.Collectors;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.cloud.client.ServiceInstance;
+import org.springframework.cloud.client.discovery.DiscoveryClient;
+import org.springframework.core.ParameterizedTypeReference;
+import org.springframework.http.HttpMethod;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.springframework.web.client.RestTemplate;
 
 import com.example.transaction.custom_exceptions.CustomException;
 import com.example.transaction.models.CsvPortfolioReader;
+import com.example.transaction.models.Portfolio;
 import com.example.transaction.models.Transaction;
 import com.example.transaction.models.bin.GetAssetQtyOutputBin;
 import com.example.transaction.models.bin.GetTransactionAfterDateBin;
@@ -25,6 +33,9 @@ public class TransactionServiceImpl implements TransactionService {
 
 	@Autowired
 	private TransactionRepository transactionRepository;
+
+	@Autowired
+	private DiscoveryClient discoveryClient;
 
 	@Override
 	public List<Transaction> getAllTransactionsByPortfolioId(long portfolioId) {
@@ -66,6 +77,11 @@ public class TransactionServiceImpl implements TransactionService {
 			throw new CustomException("Transaction not found");
 		}
 
+		// Check if current user is the owner of the portfolio
+		if (!isOwner(transactionBin.getTransaction().getPortfolioId(), transactionBin.getUserId())) {
+			throw new CustomException("User is not the owner of the portfolio");
+		}
+
 		TransactionEntity entity = transactionRepository.save(TransactionEntity.builder()
 				.id(transactionBin.getId())
 				.type(Optional.ofNullable(
@@ -91,12 +107,31 @@ public class TransactionServiceImpl implements TransactionService {
 	}
 
 	@Override
-	public void deleteTransaction(long id) {
+	public Transaction deleteTransaction(long id, String uid) {
+		Transaction deletedTransaction;
+		try {
+			deletedTransaction = this.getTransactionById(id);
+		} catch (CustomException e) {
+			return null;
+		}
+
+		// Check if current user is the owner of the portfolio
+		if (!isOwner(deletedTransaction.getPortfolioId(), uid)) {
+			return null;
+
+		}
+
 		transactionRepository.delete(TransactionEntity.builder().id(id).build());
+		return deletedTransaction;
 	}
 
 	@Override
 	public List<Transaction> saveTransactionsFromCsv(UploadBin bin) throws CustomException, IOException {
+		// Check if current user is the owner of the portfolio
+		if (!isOwner(bin.getPortfolioId(), bin.getUserId())) {
+			throw new CustomException("User is not the owner of the portfolio");
+		}
+
 		List<Transaction> transactions = CsvPortfolioReader.readCsvFile(bin.getInputStream());
 		Map<String, Double> assetsQty = this.getAssetsQtyByPortfolioId(bin.getPortfolioId()).stream().collect(
 				Collectors.toMap(GetAssetQtyOutputBin::getSymbolId, GetAssetQtyOutputBin::getAmount));
@@ -149,6 +184,36 @@ public class TransactionServiceImpl implements TransactionService {
 	@Override
 	public List<GetAssetQtyOutputBin> getAssetsQtyByPortfolioId(long portfolioId) {
 		return this.transactionRepository.findAssetsQtyByPortfolioId(portfolioId);
+	}
+
+	private boolean isOwner(long portfolioId, String uid) {
+		List<ServiceInstance> instances = discoveryClient.getInstances("portfolio-service");
+
+		if (instances != null && !instances.isEmpty()) {
+
+			// Get a random instance
+			String url = instances.get(new Random().nextInt(instances.size())).getUri().toString();
+			RestTemplate restTemplate = new RestTemplate();
+			try {
+				ResponseEntity<List<Portfolio>> responseEntity = restTemplate.exchange(
+						url + "/get/user/" + uid,
+						HttpMethod.GET,
+						null,
+						new ParameterizedTypeReference<List<Portfolio>>() {
+						});
+
+				List<Portfolio> portfolios = Optional.ofNullable(responseEntity.getBody()).orElse(List.of());
+
+				return portfolios.stream()
+						.anyMatch(portfolio -> portfolio.getId() == portfolioId);
+
+			} catch (Exception e) {
+				// throw new CustomException("Portfolio service is not available");
+			}
+
+		}
+
+		return false;
 	}
 
 }
